move all stuff, updates setup, update experiments.py
